projects/angular-ionic-browser/src/app/page/login-material/login-material.component.spec.ts
import { MatCardModule } from '@angular/material/card';
import { MatButtonModule } from '@angular/material/button';
import { BrowserModule } from '@angular/platform-browser';
import { TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { LoginMaterialComponent } from './login-material.component';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';

describe('LoginMaterialComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        BrowserModule,
        MatButtonModule,
        MatCardModule,
        FormsModule,
        ReactiveFormsModule,
        BrowserAnimationsModule,
        RouterTestingModule
      ],
      declarations: [
        LoginMaterialComponent
      ],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(LoginMaterialComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'login-material-angular'`, () => {
    const fixture = TestBed.createComponent(LoginMaterialComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('login-material-angular');
    expect(app.iconEmail).toEqual('perm_identity');
    expect(app.iconPassword).toEqual('lock');
  });
});
